Improved javascript codes and code formatting
There were several changes that were required to use the new ES standards. For example limiting the scope of the variable by using let instead of var. anonymous functions were replaced by lambda operator. To avoid linting errors in future the codes were formatted. A clearInterval() was introduced to stop the timer from going endless as it was not needed after it hits the new year date.